refactor: split useSettings into individual hooks, move validators to feature dir
- Split useSettings.ts (112 lines, 3 concerns) into:
  - useChangePassword.ts — password change logic
  - useDeleteAccount.ts — account deletion logic
  - useBlockList.ts — block list management
- Extract shared async action pattern into hooks/useAsyncAction.ts
- Move settingsValidators.ts from lib/ to features/settings/validators.ts
- Split monolithic test file into per-hook test files
- Update all component imports to use individual hooks
- useSettings.ts now re-exports for backward compatibility
